feat: HaTS survey (#5153)
This adds integration with surveymonkey.com so we can run the feedback
surveys.
There is a middleware feedback that is added in all `/docs` routes that
checks if the user should get the CTA to fill the survey. The criteria
is:
- There is a survey running
- No popup in the last 90 days
- Shows only to 8% of the users

This middleware injects some variables in the handlebars context that
are then read by the partial `feedback.hbs` that adds the right code to
show the content.